ml/requirements.txt
spacy>=3.8.11
openpyxl>=3.1.5
psycopg2-binary>=2.9.11
pandas>=2.3.3
tqdm>=4.67.1
python-dotenv>=1.2.1
